refactor: extract and expose `preloadRouteComponents`
