+++ packages/nuxt/src/app/components/nuxt-link.ts
@@ -1,7 +1,8 @@
 import { defineComponent, h, ref, resolveComponent, PropType, computed, DefineComponent, ComputedRef, onMounted, onBeforeUnmount } from 'vue'
-import type { RouteLocationRaw, Router } from 'vue-router'
+import type { RouteLocationRaw } from 'vue-router'
 import { hasProtocol } from 'ufo'
 
+import { preloadRouteComponents } from '../composables/preload'
 import { navigateTo, useRouter } from '../composables/router'
 import { useNuxtApp } from '../nuxt'
 
@@ -327,32 +328,4 @@ function isSlowConnection () {
   const cn = (navigator as any).connection as { saveData: boolean, effectiveType: string } | null
   if (cn && (cn.saveData || /2g/.test(cn.effectiveType))) { return true }
   return false
-}
-
-async function preloadRouteComponents (to: string, router: Router & { _nuxtLinkPreloaded?: Set<string>; _preloadPromises?: Array<Promise<any>> } = useRouter()): Promise<void> {
-  if (process.server) { return }
-
-  if (!router._nuxtLinkPreloaded) { router._nuxtLinkPreloaded = new Set() }
-  if (router._nuxtLinkPreloaded.has(to)) { return }
-  router._nuxtLinkPreloaded.add(to)
-
-  const promises = router._preloadPromises ||= []
-
-  if (promises.length > 4) {
-    // Defer adding new preload requests until the existing ones have resolved
-    return Promise.all(promises).then(() => preloadRouteComponents(to, router))
-  }
-
-  const components = router.resolve(to).matched
-    .map(component => component.components?.default)
-    .filter(component => typeof component === 'function')
-
-  for (const component of components) {
-    const promise = Promise.resolve((component as Function)())
-      .catch(() => {})
-      .finally(() => promises.splice(promises.indexOf(promise)))
-    promises.push(promise)
-  }
-
-  await Promise.all(promises)
 }
